% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalContour.R
\name{survivalContour}
\alias{survivalContour}
\title{survivalContour}
\usage{
survivalContour(
  data,
  model,
  contCov,
  D3 = FALSE,
  CI3D = FALSE,
  contCovName = NULL,
  nCovEval = 30,
  otherCov = NULL,
  strataName = NULL
)
}
\arguments{
\item{data}{The original data used fit the model.}

\item{model}{The model used to analyze the data.}

\item{contCov}{Specify which continuous covariate is used for the y-axis of the contour plot.}

\item{D3}{Should "survivalContour" generate a 2D contour plot, or a 3D contour plot.}

\item{CI3D}{Should "survivalContour" generate a semi-transparent layer of confidence intervals for 3D contour plot. This option is only applicable for Cox model or stratified Cox model with or without interval censored data, and parametric or spline models. Fine and Gray model for competing risks data or deep neural network models are not applicable.}

\item{contCovName}{A character string on the name of the continuous covariate shown on the contour plot. If not provided by the user, the printed name will be the covaraite name.}

\item{nCovEval}{Number of covaraite values where predicted survival is evaluated.}

\item{otherCov}{A vector of values of other covariates except for the continuous covariate used to generate contour plot. The default value is NULL, and the mean of continuous covariates and the most frequent values of categorical covariates are used.}

\item{strataName}{The print name of the stratifying covariate used in stratified Cox model. If not provided by the user, "survivalContour" will use the name of the stratifying covariate.}
}
\value{
The result of the function.
}
\description{
Show the predicted survival or cumulative incidence function (competing risks data) over time for single continuous covariate in the form of contour plot or 3D contour plot. The estimate for survival probability and cumulative incidence function(CIF) are based on various survival models. The models that survivalContour supports are listed below:
| Model | Package | Object |Confidence interval in 3D plot|
  |---------|---------|---------|---------|
  | Cox model| survival| coxph | Y |
  | stratified Cox model| survival |coxph| Y |
  | Cox model with interval censored data| mets |phreg| Y |
  | stratified Cox model with interval censored data| mets |phreg| Y |
  | parametric and spline model | flexsurv | flexsurvreg|Y |
  | Fine and Gray model for competing risks data | riskRegression+prodlim | FRG|N |
  | Fine and Gray model for competing risks data with interval censored data|intccr |ciregic| N |
  |coxtime,deepsurv,deephit,loghaz,pchazard| reticulate+survivalmodels|pycox|N|
}
